ring/tests/dev_requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o ring/tests/dev_requirements.txt --extra dev
alembic==1.14.0
    # via ring (pyproject.toml)
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==23.1.0
    # via pwdlib
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==3.0.0
    # via stack-data
bcrypt==4.2.1
    # via pwdlib
beautifulsoup4==4.12.3
    # via bs4
billiard==4.2.1
    # via celery
boto3==1.35.86
    # via ring (pyproject.toml)
boto3-stubs==1.35.86
    # via ring (pyproject.toml)
botocore==1.35.86
    # via
    #   boto3
    #   pynamodb
    #   s3transfer
botocore-stubs==1.35.86
    # via boto3-stubs
bs4==0.0.2
    # via ring (pyproject.toml)
celery==5.4.0
    # via
    #   ring (pyproject.toml)
    #   flower
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.0
    # via requests
cli-helpers==2.3.1
    # via pgcli
click==8.1.8
    # via
    #   ring (pyproject.toml)
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pgcli
    #   pgspecial
    #   rich-toolkit
    #   typer
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
configobj==5.0.9
    # via
    #   cli-helpers
    #   pgcli
cryptography==44.0.0
    # via python-jose
decorator==5.1.1
    # via ipython
distlib==0.3.9
    # via virtualenv
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.0
    # via python-jose
email-validator==2.2.0
    # via fastapi
execnet==2.1.1
    # via pytest-xdist
executing==2.1.0
    # via stack-data
factory-boy==3.3.1
    # via pytest-factoryboy
faker==33.1.0
    # via factory-boy
fastapi==0.115.6
    # via ring (pyproject.toml)
fastapi-cli==0.0.7
    # via fastapi
filelock==3.16.1
    # via virtualenv
flower==2.0.1
    # via ring (pyproject.toml)
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
humanize==4.11.0
    # via flower
identify==2.6.3
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
inflection==0.5.1
    # via pytest-factoryboy
iniconfig==2.0.0
    # via pytest
ipython==8.31.0
    # via
    #   ring (pyproject.toml)
    #   ipython-autoimport
    #   ipython-autotime
ipython-autoimport==0.5
    # via ring (pyproject.toml)
ipython-autotime==0.3.2
    # via ring (pyproject.toml)
itsdangerous==2.2.0
    # via fastapi
jedi==0.19.2
    # via ipython
jinja2==3.1.5
    # via fastapi
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.4.2
    # via celery
mako==1.3.8
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
matplotlib-inline==0.1.7
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
mypy-boto3-cloudformation==1.35.64
    # via boto3-stubs
mypy-boto3-dynamodb==1.35.74
    # via boto3-stubs
mypy-boto3-ec2==1.35.82
    # via boto3-stubs
mypy-boto3-lambda==1.35.68
    # via boto3-stubs
mypy-boto3-rds==1.35.82
    # via boto3-stubs
mypy-boto3-s3==1.35.81
    # via boto3-stubs
mypy-boto3-ses==1.35.68
    # via
    #   ring (pyproject.toml)
    #   boto3-stubs
mypy-boto3-sqs==1.35.0
    # via boto3-stubs
nodeenv==1.9.1
    # via pre-commit
orjson==3.10.12
    # via fastapi
packaging==24.2
    # via
    #   pytest
    #   pytest-factoryboy
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pgcli==4.1.0
    # via ring (pyproject.toml)
pgspecial==2.1.3
    # via pgcli
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pprintpp==0.4.0
    # via ring (pyproject.toml)
pre-commit==4.0.1
    # via ring (pyproject.toml)
prometheus-client==0.21.1
    # via flower
prompt-toolkit==3.0.48
    # via
    #   click-repl
    #   ipython
    #   pgcli
psycopg==3.2.3
    # via
    #   pgcli
    #   pgspecial
psycopg-binary==3.2.3
    # via ring (pyproject.toml)
psycopg2-binary==2.9.10
    # via ring (pyproject.toml)
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pwdlib==0.2.1
    # via ring (pyproject.toml)
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.10.4
    # via
    #   ring (pyproject.toml)
    #   fastapi
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-extra-types==2.10.1
    # via fastapi
pydantic-settings==2.7.0
    # via fastapi
pygments==2.18.0
    # via
    #   cli-helpers
    #   ipython
    #   pgcli
    #   rich
pyjwt==2.10.1
    # via ring (pyproject.toml)
pynamodb==5.5.1
    # via ring (pyproject.toml)
pytest==8.3.4
    # via
    #   ring (pyproject.toml)
    #   pytest-factoryboy
    #   pytest-pretty
    #   pytest-xdist
pytest-factoryboy==2.7.0
    # via ring (pyproject.toml)
pytest-pretty==1.2.0
    # via ring (pyproject.toml)
pytest-xdist==3.6.1
    # via ring (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   faker
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-jose==3.3.0
    # via ring (pyproject.toml)
python-multipart==0.0.20
    # via fastapi
pytz==2024.2
    # via flower
pyyaml==6.0.2
    # via
    #   fastapi
    #   pre-commit
    #   uvicorn
redis==5.2.1
    # via ring (pyproject.toml)
requests==2.32.3
    # via ring (pyproject.toml)
rich==13.9.4
    # via
    #   pytest-pretty
    #   rich-toolkit
    #   typer
rich-toolkit==0.12.0
    # via fastapi-cli
rsa==4.9
    # via python-jose
ruff==0.8.4
    # via ring (pyproject.toml)
s3transfer==0.10.4
    # via boto3
setproctitle==1.3.4
    # via pgcli
shellingham==1.5.4
    # via typer
shortuuid==1.0.13
    # via ring (pyproject.toml)
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.36
    # via
    #   ring (pyproject.toml)
    #   alembic
sqlparse==0.5.3
    # via
    #   pgcli
    #   pgspecial
stack-data==0.6.3
    # via ipython
starlette==0.41.3
    # via fastapi
tabulate==0.9.0
    # via cli-helpers
tornado==6.4.2
    # via flower
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typer==0.15.1
    # via fastapi-cli
types-awscrt==0.23.6
    # via botocore-stubs
types-s3transfer==0.10.4
    # via boto3-stubs
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   faker
    #   fastapi
    #   psycopg
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   pytest-factoryboy
    #   rich-toolkit
    #   sqlalchemy
    #   typer
tzdata==2024.2
    # via
    #   celery
    #   kombu
ujson==5.10.0
    # via fastapi
urllib3==2.2.3
    # via
    #   botocore
    #   requests
uvicorn==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.28.0
    # via pre-commit
watchdog==4.0.2
    # via ring (pyproject.toml)
watchfiles==1.0.3
    # via uvicorn
wcwidth==0.2.13
    # via
    #   prompt-toolkit
    #   tabulate
websockets==14.1
    # via uvicorn
